+++ convex/world/textures.ts
@@ -1,7 +1,8 @@
-import { ObjectType, v } from 'convex/values';
+import { ObjectType, v, ConvexError } from 'convex/values';
 import { idWorld } from '../worlds';
 import { defineTable } from "convex/server";
 import { mutation, query } from '../_generated/server';
+import { api, internal } from "../_generated/api";
 import { Doc, Id } from "../_generated/dataModel";
 import { idStorage, StorageId } from "../shared/storage"
 
@@ -10,6 +11,7 @@ export const indexName_ByWorldId = 'byWorldId';
 export const idTexture = v.id(table);
 
 export const textureSerialized = {
+  modifyTime: v.number(),
   name: v.string(),
   worldId: idWorld,
   // The speed of the animation. Can be tuned depending on the side and speed of the NPC.
@@ -18,7 +20,7 @@ export const textureSerialized = {
 };
 
 
-const { url: _url, ...insertArgs } = textureSerialized
+const { modifyTime, url: _url, ...insertArgs } = textureSerialized
 const { worldId: _, ..._updateArgs } = insertArgs
 const updateArgs = { id: idTexture, ..._updateArgs }
 const deleteArgs = { id: idTexture }
@@ -38,8 +40,9 @@ export const tableSchema = defineTable(textureSerialized)
 export const create = mutation({
   args: insertArgs,
   handler: async (ctx, args) => {
+    const modifyTime = +new Date()
     const url = await ctx.storage.getUrl(args.storageId)
-    return await ctx.db.insert(table, { ...args, url: url ?? "" });
+    return await ctx.db.insert(table, { ...args, url: url ?? "", modifyTime });
   },
 });
 
@@ -67,16 +70,23 @@ export const update = mutation({
     const { id, ...patchData } = args
     const entity = await ctx.db.get(id);
     if (!entity) {
-      throw new Error(`Invalid \`${table}\` ID: ${args.id}`);
+      throw new ConvexError(`Invalid \`${table}\` ID: ${args.id}`);
     }
+    const modifyTime = +new Date()
     if (args.storageId !== entity.storageId) {
       const url = await ctx.storage.getUrl(args.storageId);
-      await ctx.db.patch(id, { ...patchData, url: url ?? "" });
+      await ctx.db.patch(id, { ...patchData, url: url ?? "", modifyTime });
       await ctx.storage.delete(entity.storageId)
       // 删除旧的stoargeId
     } else {
-      await ctx.db.patch(id, patchData);
+      await ctx.db.patch(id, {...patchData, modifyTime});
     }
+
+    // make ref entity updateModifyTime
+    const spritesheets = await ctx.runQuery(api.world.spritesheets.listByTexture, { worldId: entity.worldId, textureId: entity._id })
+    await Promise.all(spritesheets.map(async (spritesheet) => {
+      await ctx.runMutation(internal.world.spritesheets.updateModifyTime, { id: spritesheet._id })
+    }))
   },
 });
 
@@ -87,6 +97,10 @@ export const delete_ = mutation({
     const entity = await ctx.db.get(id);
     if (!entity) {
       throw new Error(`Invalid \`${table}\` ID: ${args.id}`);
+    }
+    const spritesheets = await ctx.runQuery(api.world.spritesheets.listByTexture, { worldId: entity.worldId, textureId: entity._id })
+    if (spritesheets.length > 0) {
+      throw new ConvexError(`current texture reference by spritesheets:[${spritesheets.map(v=>v.name).join()}]`);
     }
     await ctx.db.delete(id);
     await ctx.storage.delete(entity.storageId)
